refactor(app): remove orphaned SurfaceRegistry after slot unification
The SurfaceRegistry class and its types were fully replaced by SlotRegistry
facades; SurfaceEntry moves into slot-registry.ts. Deleting the never-loaded
src/surface module fixes the packages/app coverage gate (missing 0).
